refactor: unify tag parsing
