+++ frontend/src/store/slices/deck-edits.ts
@@ -1,15 +1,11 @@
-import {
-  CardTagSchema,
-  type Id,
-  SPECIAL_CARD_CODES,
-} from "@arkham-build/shared";
+import { type Id, SPECIAL_CARD_CODES } from "@arkham-build/shared";
 import type { StateCreator } from "zustand";
 import { assert } from "@/utils/assert";
 import { cardLimit, displayAttribute } from "@/utils/card-utils";
 import { capitalize } from "@/utils/formatting";
 import { range } from "@/utils/range";
 import { clampAttachmentQuantity } from "../lib/attachments";
-import { normalizeCardTagName, resolveCardTagCardCode } from "../lib/card-tags";
+import { parseCardTagNames, resolveCardTagCardCode } from "../lib/card-tags";
 import { randomBasicWeaknessForDeck } from "../lib/random-basic-weakness";
 import { getDeckLimitOverride } from "../lib/resolve-deck";
 import { dehydrate } from "../persist";
@@ -369,7 +365,7 @@ export const createDeckEditsSlice: StateCreator<
         selectLookupTables(state).relations.fronts,
         cardCode,
       );
-      const nextTagNames = parseDeckCardTagNames(tagNames);
+      const nextTagNames = parseCardTagNames(tagNames);
       const edits = currentEdits(state, deckId);
 
       return {
@@ -536,22 +532,6 @@ function getCardQuantityUpdate(
   }
 
   return nextState;
-}
-
-function parseDeckCardTagNames(tagNames: string[]) {
-  const result: string[] = [];
-  const seen = new Set<string>();
-
-  for (const name of tagNames) {
-    const tagName = CardTagSchema.parse(name);
-    const normalizedName = normalizeCardTagName(tagName);
-    if (seen.has(normalizedName)) continue;
-
-    seen.add(normalizedName);
-    result.push(tagName);
-  }
-
-  return result;
 }
 
 function getCardUpgrade(
